1-Archive/requirements.txt
CacheControl==0.12.5
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
click==6.7
firebase-admin==2.16.0
Flask==1.0.2
gcloud==0.17.0
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.3
grpcio==1.19.0
gunicorn==19.7.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.8
itsdangerous==0.24
Jinja2==2.10
jws==0.1.3
logger==1.4
MarkupSafe==1.0
msgpack==0.6.1
numpy==1.17.2
oauth2client==3.0.0
pandas==1.0.1
Pillow==5.2.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.7.0
rsa==3.4.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.11.0
sklearn==0.0
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
XlsxWriter==1.0.9
